+++ repo_people/__init__.py
@@ -2,4 +2,4 @@ from .repo_people import RepoPeople, UserDataView
 
 __all__ = ["RepoPeople", "UserDataView", "__version__"]
 
-__version__ = "1.1.0"
+__version__ = "1.1.1"